Make types used in evm testcore match wasm types

+++ packages/evm/evmtest/TestCore.sol
@@ -7,7 +7,7 @@ import "@isccontract/ISC.sol";
 
 contract TestCore {
 
-    function fibonacci(uint32 n) public pure returns(uint32) {
+    function fibonacci(uint64 n) public pure returns(uint64) {
         if (n == 0 || n == 1) {
             return n;
         }
@@ -15,26 +15,26 @@ contract TestCore {
         return fibonacci(n - 1) + fibonacci(n - 2);
     }
 
-    function fibonacciIndirect(uint32 n) public view returns(uint32) {
+    function fibonacciIndirect(uint64 n) public view returns(uint64) {
         if (n == 0 || n == 1) {
             return n;
         }
-        uint32 n1;
-        uint32 n2;
+        uint64 n1;
+        uint64 n2;
         n1 = TestCore(this).fibonacciIndirect(n - 1);
         n2 = TestCore(this).fibonacciIndirect(n - 2);
         return n1 + n2;
     }
 
-    event FibonacciResultEvent(uint32 n);
-    function fibonacciLoop(uint32 n) public returns(uint32) {
+    event FibonacciResultEvent(uint64 n);
+    function fibonacciLoop(uint64 n) public returns(uint64) {
         if (n == 0) {
             return 0;
         }
-        uint32 a = 1;
-        uint32 b = 1;
-        for (uint32 i = 2;i < n;i++) {
-            uint32 c = a + b;
+        uint64 a = 1;
+        uint64 b = 1;
+        for (uint64 i = 2;i < n;i++) {
+            uint64 c = a + b;
             a = b;
             b = c;
         }
@@ -42,8 +42,8 @@ contract TestCore {
         return b;
     }
 
-    function loop(uint32 n) public pure {
-        for (uint32 i = 0;i < n;i++) {
+    function loop(uint64 n) public pure {
+        for (uint64 i = 0;i < n;i++) {
             // do nothing just burn gas
         }
     }
