+++ packages/vm/core/evm/evmtest/utils_test.go
@@ -439,22 +439,22 @@ func (l *loopContractInstance) loop(opts ...ethCallOptions) (res callFnResult, e
 	return l.callFn(opts, "loop")
 }
 
-func (t *testCoreContractInstance) fibonacci(n uint32, opts ...ethCallOptions) (res callFnResult, err error) {
+func (t *testCoreContractInstance) fibonacci(n uint64, opts ...ethCallOptions) (res callFnResult, err error) {
 	return t.callFn(opts, "fibonacci", n)
 }
 
-func (t *testCoreContractInstance) fibonacciIndirect(n uint32, opts ...ethCallOptions) (res callFnResult, err error) {
+func (t *testCoreContractInstance) fibonacciIndirect(n uint64, opts ...ethCallOptions) (res callFnResult, err error) {
 	return t.callFn(opts, "fibonacciIndirect", n)
 }
 
-func (t *testCoreContractInstance) fibonacciLoop(n uint32, v interface{}, opts ...ethCallOptions) (res callFnResult, err error) {
+func (t *testCoreContractInstance) fibonacciLoop(n uint64, v interface{}, opts ...ethCallOptions) (res callFnResult, err error) {
 	res, err = t.callFn(opts, "fibonacciLoop", n)
 	err = t.abi.UnpackIntoInterface(v, "FibonacciResultEvent", res.evmReceipt.Logs[0].Data)
 	require.NoError(t.chain.t, err)
 	return
 }
 
-func (t *testCoreContractInstance) loop(n uint32, opts ...ethCallOptions) (res callFnResult, err error) {
+func (t *testCoreContractInstance) loop(n uint64, opts ...ethCallOptions) (res callFnResult, err error) {
 	return t.callFn(opts, "loop", n)
 }
